job_2/requirements.txt
google-cloud-storage==2.7.0
pandas==1.5.2
pyarrow==10.0.1
pyspark==3.2.3